+++ MainWindow/FeatureDialog.cpp
@@ -1,14 +1,25 @@
-// SPDX-FileCopyrightText: 2003-2020 The KPhotoAlbum Development Team
-// SPDX-FileCopyrightText: 2021 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
-// SPDX-FileCopyrightText: 2022 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
+// SPDX-FileCopyrightText: 2005-2022 Jesper K. Pedersen <jesper.pedersen@kdab.com>
+// SPDX-FileCopyrightText: 2007 Dirk Mueller <mueller@kde.org>
+// SPDX-FileCopyrightText: 2007 Jan Kundrát <jkt@flaska.net>
+// SPDX-FileCopyrightText: 2007-2009 Laurent Montel <montel@kde.org>
+// SPDX-FileCopyrightText: 2007-2009 Tuomas Suutari <tuomas@nepnep.net>
+// SPDX-FileCopyrightText: 2008 Albert Astals Cid <aacid@kde.org>
+// SPDX-FileCopyrightText: 2009 Andrew Coles <andrew.i.coles@googlemail.com>
+// SPDX-FileCopyrightText: 2009-2016 Yuri Chornoivan <yurchor@ukr.net>
+// SPDX-FileCopyrightText: 2012 Miika Turkia <miika.turkia@gmail.com>
+// SPDX-FileCopyrightText: 2012-2024 Johannes Zarl-Zierl <johannes@zarl-zierl.at>
+// SPDX-FileCopyrightText: 2013 Pino Toscano <pino@kde.org>
+// SPDX-FileCopyrightText: 2014-2019 Tobias Leupold <tl@stonemx.de>
+// SPDX-FileCopyrightText: 2016 Luigi Toscano <luigi.toscano@tiscali.it>
+// SPDX-FileCopyrightText: 2018 Antoni Bella Pérez <antonibella5@yahoo.com>
 //
 // SPDX-License-Identifier: GPL-2.0-or-later
 
 #include "FeatureDialog.h"
-#include <config-kpa-videobackends.h>
 
 #include <kpabase/config-kpa-marble.h>
 #include <kpabase/config-kpa-plugins.h>
+#include <kpabase/config-kpa-videobackends.h>
 #include <kpaexif/Database.h>
 
 #include <KLocalizedString>